nsdiff: fix SOA regex to allow unusual syntax
The SOA RNAME field is not a hostname. Allow backslash-quoted
metacharacters, like the example in section 5.3 of RCF 1035.

The hostname regex no longer accepts address records at the root apex.

Zone names do not have to be hostnames, for example, RFC 2317
classless reverse DNS zones with / in the name, or leaf zones
containing SOA records at the apex.

Reported-by: Jordan Rieger <jordan@webnames.ca>