nsvi: tell nsdiff to get the old zone from localhost by default
But nsvi is a bit different because it assumes by default it is
running on the update server. And it also assumes that it will work to
use nsdiff's default mode of diffing against the current state of the
DNS. So, if the user has a test zone disconnedted from the DNS, nsvi
can easily get into a situation where it presents them with a zone
file to edit, and later runs nsdiff in a way that can't work.

So when the user doesn't provide a -s option, tell nsdiff to get the
old zone from localhost same as we did before the edit, which also
matches the way we invoke nsupdate.

Reported-by: Petr Menšík <pemensik@redhat.com>