nsdiff: only bump the serial number when required
The original intent of nsdiff was that serial numbers are managed
automatically, so they are basically ignored. But I changed it to
alow users to just bump the serial number - except this only
makes sense in the (default) -Sfile mode. Ignoring serial numbers
completely makes more sense in -Sunix mode (etc.)

Reported-by: Martin Hartl <mhartl@barracuda.com>